Allow launch script to start any combination of servers
Currently, the launch options are only all, api, engine, and
executor. We want to allow launch options to start any combination
of servers such as api and engine or engine and executor. The
launch script is modified to the --server command line option to
accept a comma delimited list.

Change-Id: I4ce31f5fabd4de3b2f8f0e4c5ef661ed4ec7e5b2
Implements: blueprint mistral-launch-options