Make executor able to work in isolated environment
The idea is to simplify executor as much as possible (first of all,
remove all db calls). As of now, all the work related to creating,
resolving and updating tasks, converting action params and results and
so on should be done by engine. Executer here is barely an action
runner that receives an action name, params and some kind of session
identifier (task id in our case) and reports back the status and
results along with the session identifier to reference the task it
have done.

In order to keep the model simple, we had to make several assumptions:
 - Any task that was queued for execution is expected to be executed
 and eventually report back by putting its results into
 `convey_task_result` queue.
 - All executors are the same. Each of them should be able to execute
 the action and return the same result.
 - All executor should be also able to access any external resource
 without any additional params evaluation on an executor side

Conflicts:
	mistral/engine/drivers/default/executor.py

Change-Id: I4f1f8fb08cd977ba90f69462108e15f9cfb26250