Implement workflow execution environment - part 2
Modify engine start_workflow to take either a named environment
or adhoc environment (dict) in params. Save a copy of the
environment in start_params of the execution DB object. Subworkflows
will inherit environment from parent workflow. This is part 2 of a
series to implement execution environment.

Change-Id: I9af8dbf0f9ab98de81fff5612116203beae65c93
Implements: blueprint mistral-execution-environment