Fix list of upstream tasks for task with no join
Currently, if a task is referenced multiple times in the workflow,
the upstream tasks identified are all the tasks that reference it.
However, the upstream task should only be the immediate parent if
there is no join in the task spec. If the task has a join, then the
upstream tasks should only be all the immediate parents that end
with the join and not every parent tasks that reference it. This
patch only fixes task with no join.

Change-Id: Iacf7cf5da3c91dea5412728624146873092d3686
Closes-Bug: 1427933