Add Service API
With this patch, users can retrieve all the service names via
v2/services. For example:

$ http GET http://127.0.0.1:8989/v2/services
HTTP/1.0 200 OK
{
    "services": [
        {
	    "name": "host1_6277",
	    "type": "api_group"
        },
	{
	    "name": "host2_3422",
	    "type": "engine_group"
	},
	{
	    "name": "host3_4567",
	    "type": "executor_group"
	}
    ]
}

NOTE: the Service API feature is optional for Mistral for the time being,
it's up to the cloud operator to decide whether to support it by
pre-installing zookeeper or other tooz backend component and make the
configuration correct at the same time. Otherwise, an exception will be
raised when users call the service REST API.

Change-Id: I84c882fe9c2401fe259e363428bc3c5bea0871c6
Implements: blueprint mistral-service-api