Add fields filter for workflow query
With this patch, users could retrieve workflows only with fields they care
about.

NOTE: 'id' must be included in fields if it's provided, since 'id' will be
used when constructing 'next' link.

This feature is just like we specicy '-c' argument in CLI, but it's
implemented in server side(or more precisely, in db layer).

Partially-Implements: blueprint query-result-column-filtering
Change-Id: If034e7ee96aaecf2d5cebefdc85eaa7b3f6f7d8c