Refactor get_task_spec using mechanism of polymorphic DSL entities
* Changing the way of getting task spec using polymorphic mechanism,
which is the motivation of this patch.

* Moving the type injection in TaskSpec to WorkflowSpec, the 'type'
param is already there when TaskSpec initializing, and we could also
continue to use that for polymorphic mechanism of TaskSpec, and the
subclasses iteration mechanism can also be removed as well.

* Improving the schema validation for TaskSpec, replace the code logic
of prohibition of co-existance between action and workflow in task spec
with jsonschema restriction.

Change-Id: Ia88ce390298d54afc1ee72ebfb67de9056d9f444