Add decode() function for string comparison.
A prefix of 'b' or 'B' is ignored in Python 2
But in py3 produce an instance of the bytes
type instead of the string type.

Due to which it gives following kind of error.
e.g.testtools.matchers._impl.MismatchError:
'unable to convert to int' not in
b'{"faultcode": "Client", "debuginfo": null,
"faultstring": "Invalid input for field/attribute limit.
Value: \'1.1\'. unable to convert to int"}'

Change-Id: I226e8f8c6afbbc94c6348a17099f02b9cf7907a8
Partially-Implements: blueprint mistral-py3