Division result is wrapped to int().
In py27 sys.getsizeof(str(long_string)) / 1024 returns 
int value but in py34 it returns float value.

Due to this "if size_kb > size_limit_kb:" condition is
getting true and in py34  UT
"Size of 'input' is 0KB which exceeds the limit of 0KB"
error is occurred.

Partially-Implements: blueprint mistral-py3

Change-Id: I0c7a4bc316fcc67084d81f59cb1f6f3f9d1a8eef