Fix state change on exception during task state change
If there is an exception during on_task_state_change, particularly if
there is a YAQL evaluation error in the input of the next task, the state
of the current task and related action execution will be marked ERROR for
sync task and rolled back for async task, leaving the task as orphaned in
RUNNING state in most cases. This is a workflow handling error and the
workflow execution state is already marked ERROR. The state of the current
task execution should not be affected.

Change-Id: Id28b0661c09efa80a666d9eb7606fb435641a989
Closes-Bug: #1510738
(cherry picked from commit 93b19335f5992f74ac0edcd4f95c0ad06d101b74)