Make test_expiration_policy_for_executions stable
mistral.tests.unit.services.test_expired_executions_policy
ExpirationPolicyTest.test_expiration_policy_for_executions
occasionally fails - depending on the runtime length of the
unit tests.

The fix is to move the initialization of WF_EXECS (which also
initializes the created_at and updated_at properties) from the
global init section of the file into the _load_executions method
so the created_at and updated_at properties will be updated with
values that are relative to the beginning of the current test and
not the load of the python file.

Change-Id: I7f17d8086bbef05fb4e8a32fbf872c683b021ccf
Partial-Bug: #1506199