Fix join on branch error
If a workflow has multiple branches that join on a task and there is
one or more unhandled errors in upstream tasks at these branches, the
workflow execution state will be RUNNING indefinitely. In this case,
the workflow execution will never reach completion and the engine
should fail the workflow.

Change-Id: I237ed0ac481f946de1626e8d7936cfb7bf9081d6
Closes-Bug: #1472790