Fix with-items concurrency for sub-workflows
For with-items task that uses concurrency to execute subworkflows, the
remaining iterations after the first iteration will return the same item
on subworkflow completion. This bug does not affect with-items task that
executes action.

The bug is caused by a delay in the scheduling of workflow execution and
the creation of the execution record. For action execution, the record is
created within the same DB transaction prior to scheduling.

This patch creates the workflow execution record first just like for
action execution and then schedule a resume workflow.

Change-Id: Iba80068260caead9ae8f2f8f105abc5b9349db52
Closes-Bug: #1536415