Fix for not running 'on-success' task after task with 'with-items'
The following bug is caused by using READ_COMMITTED transaction level. When
2 actions are running simultanously, this is what happen in default engine
when on_action_complete is called:
1. Action1 is marked as 'Success'; engine checks if Task1 has ended (it
has not, since there is still Action2);
2. Action2 is marked as 'Success'; engine checks if Task1 has ended (it
has, both actions has been marked as 'Success');
3. Action1 enters second transaction and checks if Task1 has ended (it
has, both actions has been marked as 'Success'), so it looks for next tasks
to run - it finds Task2
4. Action2 enters second transaction and checks if Task1 has ended (it
has, both actions has been marked as 'Success'), so it looks for next tasks
to run - it finds none, because of what happens in Action1 thread.

With this change, it is not only checked with DB if task has ended in
second transaction, but it also checks, if task was ended in the previous
transaction.

Closes-Bug: 1535722

Change-Id: Ifb2cfde4a88619b7664669f2d5ca14fa47379961