Change for synchronous Mistral actions from CLI
This patch allows the Mistral actions to be run synchronously from
the Mistral CLI.

When a Mistral action is run synchronously using the Mistral CLI
'run-action' command, it does not complete successfully. This is due
to the Mistral API server handling requests on a single thread.

The 'run-action' command performs a REST POST call to the
ActionExecution RestController in the API server. That in turn calls
back into the python-mistralclient which then performs another REST
call back to the appropriate REST controller to actually run the
action requested. That call hangs since the requests are handled on
a single thread because the first POST has not completed yet.
Eventually the RPC call between the engine and the executor servers
times out, and the 'run-action' command fails.

This patch changes the Mistral API server so that requests are
handled in separate threads.

Added a new functional test to the tempest test package to test
synchronous action execution of a mistral action from within
mistral.

Change-Id: I8e06d3ef6aab4b2009a8fff4aa4d1acc118eee3f
Implements: blueprint mistral-mistral-actions