Moved CORS middleware configuration into oslo-config-generator
The default values needed for mistral's implementation of cors
middleware have been moved from paste.ini into the configuration
hooks provided by oslo.config. Furthermore, these values have been
added to the default configuration parsing. This ensures
that if a value remains unset in mistral.conf, it will be set
to use sane defaults, and that an operator modifying the
configuration file will be presented with a default set of
necessary sane headers.

Change-Id: I1174fa44f316be60fdc86a4d1032ca71c9827a25
Closes-Bug: 1551836