Fix with-items task termination when sub-workflows fail
When a with-items workflow based task is executing the sub-workflows
with concurrency, the task will stay in the RUNNING state if at least
one sub-workflow fail.

This behavior do not occurs for with-items action based tasks.
The issue is caused by a different management of the "accepted" flag
for actions than sub-workflows. A with-items task wait for the
termination of the concurrent executions by searching the "accepted"
ones. An action is set as accepted when its state change to "SUCCESS"
or "ERROR", while a sub-workflow is set to "accepted" when its state
change to "SUCCESS" only.

As for actions, this patch set the sub-workflows as accepted when fail.

Change-Id: Ie94af2a6a037d1184d127a27ea86d0752c03b3b2
Closes-Bug: #1555025