Fix get task list on YAQL error in with-items
If there is a YAQL error in with-items, the with_items_context is not
initialized and get task list from the API will failed with
"ERROR (app) 'with_items_context'". This patch fixes the data_flow
module and returns a default with_items_context dictionary if it
does not exist.

Change-Id: I2cc7d71f30dfad3122e06637a7583333cb02de6b