Update get_arg_list_as_str to skip func params
This patch updates inspect_utils.get_arg_list_as_str
so that it skips handling parameters that can't be
JSON serialized. One example of this would be a function
parameter (func=time.sleep). Some of the OpenStack
clients have functions that would be useful to expose
by simply relying on the default settings for these
parameters.

Change-Id: Id1cc8e79db3001ffec5066b2e81ec2e5119e875a
Closes-bug: #1584366