Don't create actions when attempting to update one that doesn't exist
The API called create_or_update_action_definition when doing a PUT
request. This was incorrect, as updating an action that doesn't
exist shouldn't create a new action - rather a 404 is expected.

This change adds a new create_or_update_actions method that has
has the previous behaviour and updates the action_manager.py to
use this as it required the old behaviour. The update_actions
method is then updated to use the DB API call that will error if
the action doesn't exist.

Change-Id: Iea9ce71078058224172d31e20d84746052a28b2f
Closes-Bug: 1568541