Add configuration option for endpoint type
This adds a new option in the default group called os-actions-endpoint-type.
Doing so allows removing the hard coding of the parameter passed into
novaclient. Although novaclient on the CLI will append the URL suffix
to the option [1], it seems passing in the full text (e.g., publicURL) is
most appropriate.

The environment variable OS_ACTIONS_ENDPOINT_TYPE may also be specified
instead of using the option from the configuration file.

[1]
https://github.com/openstack/python-novaclient/blob/d6ca9e2d7473f6d15a3c76bcd5a7c3cc65b49296/novaclient/shell.py#L809

Change-Id: I1ac52113193f286b6f2ee53f6221c39de91d20f1
Closes-bug: #1526300