Use actual session for ironic-inspector action population
python-ironic-inspector needs an actual session when creating the
client, since it will try to fetch and furtherly validate the given
version regardless if a version is explicitly given. This fetching
and validation do calls to keystone which require credentials. Thus,
I'm using mistral's service credentials to fetch the required info.

Change-Id: I908552d2fb8ab4a5a3593f03f92ee6a31b672034