Fix past migration scripts discrepancies
It is important that running all migration scripts from first to last
will give the expected model.

This fix is a first step. It fixes all discrepancies from previous
scripts.

There are still changes in the model that needs to have a migration
script written for them, but that is not handled now.

Change-Id: Ie4079eecef6766ed91c210ccefa058175e0f22d0