Publish/output in case of task/workflow failure
Currently it is not possible to provide any reasonable output in case of
a task or workflow failure. Implementing this would greatly simplify error
handling in workflows. This blueprint is a proposal to introduce two new
attributes, publish-on-error for tasks and output-on-error for workflows
for this purpose.

Implements: blueprint mistral-publish-on-error
Change-Id: Ib3a64971effb02390206dc6f993e772a51f8f237