Update heat actions in mapping.json
The "heat resource-mark-unhealthy" (or "openstack stack resource mark
unhealthy") command allows the user to mark a resource as failed. This
will cause Heat to create a replacement during the next stack update,
while still respecting all of the dependency relationships within the
template (so e.g. if a server gets replaced then floating IPs get moved
to the new server and so on). Therefore this is critical to being able
to use Mistral to automatically recover applications after e.g. a server
dies.

The "heat stack-environment" command for downloading the environment of
a stack is also added.

Change-Id: I7888c281db121cb99c0d18491bc0ee0bd0c7b473