Migrate mistral task_type
In order to be able to get full info from the API for old tasks,
it is needed to get the type from the spec column to the type column.

The alternative was adding a release note explaining the inconsistency.

Change-Id: I48dd286f749ceb491ab32ee15af62a8db54acdc7
(cherry picked from commit 4f07468bc8e8fc4575ecbe260ae78d2ef8c19863)