Bump Ironic API version to 1.22 when creating the Ironic client
Due to how microversioning works in the Ironic client, it is not possible
to use API features introduced in API version 1.10 and later from Mistral,
as the default version is 1.9 (mid-Liberty). This change bumps the required
API version to 1.22 (Newton final).

See the following link for the full API version history:
http://docs.openstack.org/developer/ironic/dev/webapi-version-history.html

Change-Id: I1c605fc00efe5fe8956d6547ea5e85e6e1172c9b