Allow "version" to be within workflow names in workbooks
There were two bugs in the schema validation for Mistral workflow names
within Workbooks. The goal in this part of the validation is to detect the
version property and assign it the enum type and then assin all other
single word properties to use the workflow/action schemas.

1. Detecting version properties. The code before this patch has the
following line.

    "version": {"enum": ["2.0", 2.0]},

This looked safe, but "version" is actually a regular expression.
Meaning that it will match any string containging "version". Adding ^ at the
start and $ at the end means that it matches a string from the start to the
end and is exactly "version".

2. Detecting workflows/actions. The current regular expression was

    "^(?!version)\w+$": _action_schema

This worked in most cases. It basically means "Any word that doesn't
start with version". So this worked, but was incorrect for workflows
called "versionworkflow" - it didn't match this when it should have.
Using ?: to create a non-capturing portion of the regular expression
solved this issue.

Any workflows that have a name that doesn't match this regular
expression. Such as those with a space (like "my worflow") were silently
ignored before this patch. The addition of additionalProperties: False
in the schema shows users an error if they use an invalid workflow name.

Closes-Bug: #1645354
Change-Id: Ib2b406a05cf15b41be075f886de77509a9da8535