Fix configuration generator
The current behavior of 'genconfig' is writing file
'mistral.conf.sample' directly to the '/etc/mistral' directory. This has
two problems: the /etc/mistral directory may and may not exist; we are
supposed to write a usable configuration file into the '/etc/mistral'
directory instead of a sample file.

This patch redirects the output path to the local 'etc' subdirectory.
Developers/users can then tune and install the configuration file based
on the sample.

Change-Id: I0c75ab9ffc9960ba12ce6f5d725c0c7bb78ef238