Fix get_next_execution_time
get_next_execution_time was incorrectly modifying it's return value
based on a conversion from localtime to UTC. However, this function
should not be doing such a conversion. It could result in multiplying
a TZ error. Croniter doesn't care about TZ. It just applies a pattern
to a given start time; therefore, it should be up to the caller to
pass the proper time.

Also, start_time is now based on utcnow() instead of localtime now().

Change-Id: Ifbd63d9e1f56085928bede22ce4f2954e1b38991
Partial-Bug: #1654218