Change Http action result content encoding
Sometimes action result is present not in "utf-8", it cases
to "UnicodeDecodeError: 'utf8' codec can't decode byte" so
need to pass proper encoding format in serialiser and deserialiser
if it is different from default, but it should be resolved in
general as a fix for another bug (#1676411).

As a hotfix for linked bug, encoding was changed in the
run Http action method directly.

Also default encoding value in the FakeHTTPResponse was changed
from 'utf8' to 'utf-8' like in official python request lib docs.

Change-Id: I0ab9ec84c0973f362b57ab82248f27768a984678
Closes-Bug: #1667415