Remove the empty using_yaql gude
The YAQL documentation has greatly improved recently, we should just
reference that rather than starting our own documentation for YAQL.

We may want to include some Mistral specific examples, if we do, we can
re-add this file when needed. Having a file that isn't included in the
documentation tree causes Sphinx to show a warning.

Change-Id: Id7d054697038be5a2e1c03854a8e8f7259e8a7e6