Rework the CLI Guide
The previous version contained badly formatted reStructuredText which
caused rendering issues. It also contained the help output for commands,
but it wasn't updated or complete.

This new document tells used how to view and use the CLI so they can get
the help output easily themselves. This removes the rendering issues and
the need to keep this page updated.

Change-Id: Ic97185035ef22a9d6d291b9f396488139e8a7e2d