Fix gate failure
The failure:
http://logs.openstack.org/45/454945/1/check/gate-mistral-devstack-dsvm-ubuntu-xenial-nv/e9daa6e/console.html#_2017-04-08_11_50_48_519100

When populate workflow actions in database, no actual authentication is
made, as a result, for Keystone client, the 'service_catalog' property
value is None. After the patch[1], the 'service_catalog' related
actions won't be inserted into the database and can not be used in
workflow definition.

This change removes 'service_catalog' related actions from the mapping
file and modify the keystone action test case accordingly.

Change-Id: I469d38cad165b57690c43db485f86e898f08de06