Don't save @property methods with other attributes
When actions are registered, Mistral saves the attributes on the class,
if one of them is a method with the @property decorator then it will
fail - these appear to be regular attributes, not methods, but are of
the object type `property`.

This change adds a type check to ignore any property objects.

Change-Id: I7de3964b74d00d147d847ec664dc8c5d0b734321