Update AdHoc Actions to support context data references
AdHoc actions can be defined using YAQL and Jinja2 expressions in the
same way as Tasks, but they could not access the associated context
data because the context was not available when the expression is
evaluated.  This patchset passes the task and workflow context into
the AdHocAction object so that the inputs can be evaluated using the
available context, and the context data will be available for
reference.

Added a test to verify that the env() works in AdHoc Actions.

Change-Id: Ib95604d3d494a443e852bc7f5eee24f398b1648c
Closes-Bug: 1690158