Update Docker README
This patch updates the docker README file to include the --link arg
to the command for launching a Mistral container to ensure the
rabbitmq and mistral containers can communicate.

This patch also adds a new section that explains how and why to
mount the Mistral source code as a volume in a running container.

Change-Id: Iec25f7e7dfda0bc8c9b38fa37b048448d73f66e4
Co-Authored-By: Brad Crochet <brad@redhat.com>