Always retrieve endpoint in mistral action
The mistral action had a strange conditional where it would only
retrieve the endpoint when trust tokens are used, and otherwise rely on
the client to do so. It seems to make more sense to do it all the time,
especially for services which uses trust tokens to talk to Mistral (eg
Heat).

Change-Id: If37162f7df59978dc511d0e920e952fb954ba10f