Centralize session creation and authorization from OS clients
Recently novaclient moved to keystoneauth. When a token is passed to it,
an instance of Token auth plugin is used to handle it. The problem is
that it tries to reauthenticate using the token. It is not possible
with trust-scoped token.

Here we centralize the session establishment for service client creation
in most actions that enables the reuse of the existing token.

Change-Id: Ibe9ee28a027e7a782adb8d8120d745259c4608da
Co-Authored-By: Andras Kovi <akovi@nokia.com>
Closes-Bug: 1690787