Use recommended function to setup auth middleware in devstack
Currently Mistral has own configuration for keystone
auth middleware, many parameters of which are deprecated [1].
It's not desired behavior and it is suggested to use recommended
devstack configuration function to prevent possible errors if
something is changed in keystone deployment in the future.

This patch fixes this situation and implements official
"configure_auth_token_middleware" function support.

[1] https://github.com/openstack/keystonemiddleware/blob/712438ebf9ee30d553f6b515e38b1c91a9ae498e/keystonemiddleware/auth_token/_auth.py#L29-L35

Change-Id: I5f50ca6b773b61c35f93488e64a1b95d3ace9c2c
Closes-bug: #1697662