Fix cron keystone calls when token is available
When a token_id is present we get

Failed to process cron trigger CronTrigger
AuthorizationFailure: Authentication cannot be scoped to multiple targets.
Pick one of: project, domain, trust or unscoped

We have a test to do that if CONF.keystone_authtoken.auth_type is not None
but a change was needed when it was none.

Change-Id: I636873c0db4b3dbf66a0c5a856fee4dcb644ac3c
Closes-bug: #1711069