Fix services launcher to handle shutdown properly
When Mistral services are configured with systemd, the services time out
on shutdown. The root cause is that each of the server thread is
assigned to a different instance of the oslo.service ServiceLauncher.
The ServiceLauncher can manage multiple server threads. When the server
threads are added to the same instance of the ServiceLauncher, SIGTERM
is handled properly for each server threads.

Change-Id: I8b739844f6073890324189aee028c0d7da3cc76e
Closes-Bug: #1714351
(cherry picked from commit 8a4fe7f146446201d88237804e007c8987814622)