Mistral fails on RabbitMQ restart
Turns on the 'confirmation' for message publishing in the
Kombu RPC client.

Fixes the a race condition in the Kombu RPC client between the
reply queue being declared and the listener being started and
the reply being sent by the server side.

Fixes the Kombu RPC server not resetting the sleep timer after
successful connection to the MQ service.

Change-Id: I0db1cb4c2de7f2c7415825b28e961076870038bf
Closes-Bug: 1718883