Enable eventlet monkey patching for MySQLdb driver
Using MySQLdb driver for MySQL is not compatible with
greenthreads provided by eventlet library. If we issue
an SQL query through it from within a green thread then
eventlet dispatches doesn't switch to another green
thread although it is an IO operation.

We need to apply monkey patching to MySQLdb explicitly
as it was fixed some time ago but doesn't apply by default.

Change-Id: Id32c65b628d8ffec5452cc89d7cf5a0a24f0312d
Closes-bug: #1721218