Invoke AuthHook before ContextHook
Now in Mistral there is an AuthHook that should work for both Keystone
and Keycloak. But for Keystone the real call to the server is performed
earlier, in access_control [1], and AuthHook just contains a small check
of X-Identity-Status header, that was injected before.

Therefore, the sequence for Keystone auth is next:
1. Call Keystone auth middleware during access control check and inject
   required headers in the request object.
2. Call ContextHook and build the context.
3. Call Keystone AuthHook and validate X-Identity-Status header we got
   on step 1. [2]

Because of such a blurry logic, it seems that everything works as it
should.
But for Keycloak we honestly do autorization in AuthHook (where it
supposed to be), and we have the following sequence of actions:

1. Build an empty context without project_id and roles using ContextHook.
2. Call AuthHook and perform the authorization. [3]

So, even if the authorization was good and the headers were injected,
the context had already been created.

To fix this bug it's enough just to swap the order of calling the hooks,
because this is logical to build user context after authorization
process, because AuthHook may add new identity headers which can be
requred for the context builder.

[1] https://github.com/openstack/mistral/blob/master/mistral/api/access_control.py#L39
[2] https://github.com/openstack/mistral/blob/master/mistral/auth/keystone.py#L33-L35
[3] https://github.com/openstack/mistral/blob/master/mistral/auth/keycloak.py#L81

Change-Id: I33ae6ddfe2a62851401305ff1dd8890b6937e842
Closes-bug: #1719819
(cherry picked from commit 75e66117f049e00ba682f1d8977cdaa6eae35ef0)